reorganize charity-dao.ts and charity-router.ts
